refactor(lexer): use reentrant yylex
Add context manager methods to lexer classes.

BREAK CHANGES:
Remove `old_klvm` module.
Remove `state` property in all lexer classes.
Remove `KoiLang.parse_command` and `KoiLang.parse_args`.

+++ kola/unicode_handler.c
@@ -22,7 +22,7 @@ PyObject* decode_escapes(const char* s, Py_ssize_t len) {
     PyObject *v = NULL;
 
     /* check for integer overflow */
-    if (len > SIZE_MAX / 6) {
+    if ((uint64_t)len > SIZE_MAX / 6) {
         return NULL;
     }
     /* "ä" (2 bytes) may become "\U000000E4" (10 bytes), or 1:5
